+++ src/api/definitions/mod.rs
@@ -457,12 +457,7 @@ impl Module {
                     !f.metadata.name.contains('$') && !is_valid_function_name(&f.metadata.name);
 
                 #[cfg(not(feature = "no_custom_syntax"))]
-                let operator = operator
-                    || def
-                        .engine
-                        .custom_keywords
-                        .as_ref()
-                        .map_or(false, |m| m.contains_key(f.metadata.name.as_str()));
+                let operator = operator || def.engine.is_custom_keyword(f.metadata.name.as_str());
 
                 f.write_definition(writer, def, operator)?;
             }